First complete draft of getting started
I wrote a lot!!! The article covers the basics of how to get started.
It's pretty long, so I don't mind chunking it up - but I think it's nice
to have it all in one place so you can have all the relevant info in one
place. Let's chat about content strategy soon.